Add GrpcSession, so transports can expose data (client IP, certificate)
The transport creates a GrpcSession for each new client connection,
and can expose additional metadata, such as the client IP address.

GrpcSession is included in Metadata.Headers so that ServerInterceptors
can access it, for example to include the client's IP address in logs.

GrpcSession is also set as a ThreadLocal accessed through GrpcSession.get(),
so that method implementations can also get access should they need it.

Closes #174