Prepare codebase for per-plaform implementations.
Introduce Platform and Base64Codec interfaces.

Implement a default platform implementation that gives a simple
default codec implementation that just calls through to the base64
implementation in Guava.

Introduce a call site in PlatformProvider that gives the correct
platform object for the current platform (different versions of this
file will, in the future, be swapped in for different platforms).

Update TransportFrameUtil to get its base64 code from the platform
object rather than calling Guava directly so that we can later configure
PlatformProvider via the osdetector mechanism in Gradle to give us a
implementation specific for a particular platform.