POC: Separating outbound/inbound stream state
The stream class structure has become a bit complicated and hard to understand. Much of this is due to the fact that we have so many special cases for client/server, okhttp/netty, etc.

I think that we can begin to simplify the code a bit by pulling out the core logic that compose the inbound and outbound state of the streams into specialized classes.  This PR is a first-cut at what this might look like.