Remove the use of MoreObjects/Objects from CallOptions.
Guava is renaming its Objects to MoreObjects because Java 7 also has a
class named Objects. Such migration is happening inside Google. Although
gRPC depends on Guava 18, Google internally was stuck using Guava 14.0
for legacy reasons (#879) so gRPC cannot migrate to MoreObjects which is
not available in Guava 14, and it should not stick to Objects because
this would conflict with the internal migration work. To make our lives
easier, we could just drop the use of MoreObjects/Objects.