Run createTransportRunnable outside of lock.
Because `scheduleConnection()` is run under lock, if we ran
`createTransportRunnable` inside `scheduleConnection()`,
`savedDelayedTransport.setTransport()` will be under lock which violates
the assumption made in https://github.com/grpc/grpc-java/issues/1408 that

> there is an implicit rule today that channel layer will not hold any lock while calling into transport

and had caused deadlock with `InProcessTransport`.

This is a partial back-port of 0e14516f5a5cbee5e60e467bfbc48a8a3197ff40
from master. A large portion of changes on tests is not ported because
of conflict caused by API changes in master.