netty: recycle command objects on the WriteQueue
WriteQueue may not be long for this world, but we have to live with it while it is here.

This change makes it so that QueueCommand objects are recycled from a pool of "UnionCommand" objects, which represent all possible commands.  They all share a single small set of variables.  Each command is actually a tagging interface with the appropriate getters.

Because each of these objects was destined to be sent across a thread, there is basically no chance they could be stack or TLAB allocated.  I suspect the objects will flow from the application thread to the network thread, and hope that Recycler steals chunks of them back over as needed.

I don't expect this to have a dramatic impact, but it should reduce some of the noise in the allocation profiler.  By my count it saves about 2% of object allocation.

Some nice things about this:

- The if-else chain in the Handlers is gone!  It is replaced by an enum switch
- Because Union command implements all command interfaces, it can be passed to other functions without casting
- The other functions always ask for the subtype, so they don't call the wrong method.
- All commands now implement equals automatically.
- SendGrpcFrameCommand was previous in a different part of the hierarchy because it extended ByteBufHolder, and couldn't extend AbstractCommand.  The hierarchy is now uniform