netty: use void promises and improve error handling
Changed the write queue to use void promises by default, this brings the advantage that any
failures will be propagated via the channel pipeline automatically and so we don't need to
add a listener checking for errors to every write.

Additionally, improved error reporting in places where we added a listener, but did not check
for errors.