Trying to fix AbstractInteropTest flakiness
The server-side metric can be recorded *after* the RPC call has returned
on the client-side.  If the test case that doesn't check the metrics is
followed by a test caset that does, the metrics from the former may be
recorded after the latter starts, messing with its results.

The fix is to add a rolloverRecords() method in
FakeCensusContextFactory, which disconnects the factory with the record
sink, and switches to a new sink, thus the metrics from the former test
case will not be visible from the latter test case.