netty,okhttp: Fix CONNECT and its error handling
CONNECT was previously non-functioning. It is now reasonably complete.
However, both Netty and OkHttp should be broken with TLS authority
override.

In Netty, the closeFuture() closes too early to be helpful, which is
very unfortunate. We'll just need to rely on the handlers to report the
shutdown without any back-up. The handlers error handling has matured,
so maybe this is okay.

WORKING: two tests fail; both are flaky

Fixes #2193