context: expose Ticker interface and testing method
The motivation is to make it easier for users to test their
deadlines.  Currently, only packages in io.grpc can set up tests
to use a fake ticker, while all other packages have to rely on the
System ticker.  This makes tests that compare deadlines flaky.

This exposes an interface in the context package for testing.
The previous abstract class was made to match the Guava version,
so for outside users to be able to continue using Guava, it
cannot be an abstract class.

The `baseInstant` has been left private, since I think it is still
easy to test deadlines without having access to it.

Fixes #2531