testing: fix flaky okhttp test newStream_duringShutdown
resolves #2328
Root cause analysis: for OkHttp, there was a race in the test. Because the AsyncFrameWriter
of OkHttpClientTransport executes its IO tasks by a serializingExecutor,
asynchronously, it may not yet start doing any IO even when client.shutdown() is called.
If client.shutdown() happens prior to the frameWriter doing anything, then
the server does not receive the first stream, and then the test fails.