netty:  use channel writability as signal to flush
The goal here is to flush when the channel is going to buffer.  In
order to accomplish this, the channel's isWritable() method is used
to check if the state changes after writes.

The tests have been modified to not use mocks, and more accurately
check the desired behavior.  That said, the tests do not yet
reflect the target behavior.