core: set "sampled for local span" per MethodDescriptor.
This moves away from the global String-based Span name registry which
is not as flexible as we desire.

Also renamed the option name to be more accurate.  This is not
API-breaking because the origianl addition to MethodDescriptor and
code-gen didn't make it into the 1.7.0 release.

This is using a snapshot of opencensus, and needs to switch to a
release before merging.