protobuf-lite: add mark support to ProtoInputStream
This will be used for binary logging, where we want the read the first
N bytes of the stream (where N is possibly all the
bytes). ProtoInputStream is backed by a ByteArrayInputStream which
supports mark/reset. This lets us avoid creating an extra buffer to
store the bytes we read.