core,netty: correctly count streams that ended due to client cancel
Intuitively we would want to bump this counter in
AbstractServerStream.close(), the place where the application passed
in the status. But that code exeutes on the application thread and we
want to avoid synchronization. It is suboptimal that we have two far
away places that bump the counter: 1) here, 2)
NettyServerHandler.sendResponseHeaders, but these are the two most
obvious places that are serialized in the transport thread.