okhttp,netty: move StreamIdHolder to internal
This makes streamId holder common between okhttp and netty.  In a
subsequent PR, I am going to make it used by ClientCallImpl callbacks
as a debugging identifier in toString()

The reason is that sometimes a callback throws an exception and is
logged, but the client call is still alive for some reason.  It's nice
to be able to associate an exception from a callback with a particular
RPC.

One other note: this makes the `id` in Okhttp non volatile.  I looked
at all the accesses, and think this was a hold over from before
the transport state change.