test: de-flake Census tag checking in AbstractInteropTest.
Tests in AbstractInteropTest share the same server instance.  Test
methods, being on the client-side, have no way to reliably tell when a
server-side end-of-RPC record is registered into Census.  Therefore,
we try polling the server records until the expected record is found,
or there is no more record.

When I made the change to record start-of-RPCs to Census, I updated
this test to also check for such records.  I made an assumption that
the end-of-RPC record of an RPC immediately followed the corresponding
start-of-RPC record, which is not always the case.  Thus the flakiness.

Before the fix, Http2NettyLocalChannelTest fails 18 out of 1000 runs
when run internally.  After the fix, there is no more failure.