core: Use Proxy instead of ProxyDetector to specifying proxy for channel
With this we use the standard java.net.Proxy to specify proxy
information instead of exposing the internal ProxyDetector. In order to
avoid exposing internal interfaces, we also modify NameResolverProvider so
that we can evaluate whether to use a proxy before calling into the
provider. This means that the only addition to the public interface is
an overload with a boolean indicating whether a proxy should be used,
which is sufficient for what DNSNameResolver currently does.