netty: pass full authority to SSLEngine
Previously, attempting to call overrideAuthority with a valid
authority would drop the beginning portion of the authority if it
contained a User.  For example `TestUser@testserver`.   This would
make the SSL handshake fail because the certificate name woudl be
`TestUser@testserver`, but the name to match would be `testserver`.

This usage is uncommon, but it may not be possible to modify the
certificate being used.  The only previous recourse was to disable
hostname verification, which is less than ideal.

Instead, pass the authority through without the port, rather than
the hostname.  This also solves the problem with invalid host names
like "test_cert".   This is the same behavior as the Go gRPC
implementation.

Note: the SSL engine has called this parameter "host", but I
believe this is a historical artifact.