RELEASING.md: Push branch before tag
If the release branch (i.e., v1.11.x) was not up-to-date with the local
repo, then the push to update the branch will fail. Doing it first
allows you to notice, merge in the updates and fix the tag before
pushing it.

Also document pulling the most up-to-date version of the release branch
before creating the 'release' branch. That would reduce the amount of
time the release branch could be modified.