services: use opaque pagination tokens in channelz
It turns out, dealing with int64 in javascript is painful. It is
also nice not to depend on proto in the js, which means the stats
hook should convert everything into JSON.

I considered letting clients pass back in the last result from
the page as is, and let the server figure out whether to +1 to
it. But this leads edge cases that are tricky to deal with.