services,core: simplify CallId generation
Function changes after vs before:

After
=====

BinaryLog.java is the class that is responsible for intercepting
client and server calls. It now requires a CallId to be passed
in. The BinaryLogProvider is responsible for generating a
CallId.

BinaryLogProviderImpl will generate a CallId based on census info
by default. For the server CallId, my intention is to read the
census info from the Context. For the client CallId, the census
span is not accessible, so a pre-computed CallId is stored into
the CallOptions.

Before
======

The idea was to put a pre-computed CallId into the Context for
the server context, and a pre-computed CallId into the
CallOptions. But as we discussed there's some issues with how it
was trying to access the census info.