services: encode the Status in the binarylog metadata object
One of the design goals of bin log is to do what a user application
can already do, which implies we stick to public APIs and avoid low
level internals. This does not exactly match the Java API which tries
really hard to hide the h2 implementation details. If we ask the C
folks to denormalize it out to a separate proto field then it won't
quite match the C API. Let's just make Java log something first and
revisit later.